test/requirements.txt
python-keystoneclient>=1.1.0
python-novaclient>=2.22.0
python-saharaclient>=0.9.0

